requirements.txt
streamlit
pyrebase4
easyocr
roboflow
Pillow
firebase_admin
datetime
pythainlp